admission: split the main and implementation
