+++ docker/Dockerfile.admission-webhook
@@ -10,7 +10,7 @@ WORKDIR /root/networkservicemesh/
 RUN ./go-mod-download.sh
 
 ADD [".","/root/networkservicemesh"]
-RUN CGO_ENABLED=0 GOOS=linux go build -ldflags '-extldflags "-static"' -o /go/bin/admission-webhook ./k8s/cmd/admission-webhook/main.go
+RUN CGO_ENABLED=0 GOOS=linux go build -ldflags '-extldflags "-static"' -o /go/bin/admission-webhook ./k8s/cmd/admission-webhook/...
 
 FROM alpine as runtime
 COPY --from=build /go/bin/admission-webhook /bin/admission-webhook
